scanner/requirements.txt
fastapi
httpx
pydantic
python-dotenv
uvicorn
google-genai
PyJWT
bcrypt
passlib
python-multipart
